Component Class Name Refactor
Was still not happy with the duplicated code, and inconsistency of class names accross elements.

Additionally; whilst lost of class names are available, I believe this is a better option than attempting to manage a method to enable them in the component. You can just ignore them if you don't need them.

Refactored the classNames generation to maintain consistency.

Reworked tests to expect these changes.

These are, by no means; final. But they are a huge improvement.

Added examples for classnames, which demonstrates these overrides (with an array, which results in lots of classes by design).