Fix Validation Bug
Found I had handled the switch (type) of input in the same way for each type.
I have left the switch, as I know it will be needed on checkbox types (and others), but removed duplicate code.

Additionally; reset state of the base validationErrorMessage to ''.

This requires a minor fix as part of another issue to provide default validation failing onComponentDidMount.